Restructure to split browser logic from core logic
