+++ browser/game/index.ts
@@ -1,0 +1,3 @@
+export * from './game-local-2p'
+export * from './game-local-ai'
+export * from './game-online-2p'
