+++ core/board/index.ts
@@ -1,2 +1,1 @@
 export * from './base'
-export * from './ui'
