+++ core/game/index.ts
@@ -1,4 +1,2 @@
 export * from './game-base'
-export * from './game-local-2p'
-export * from './game-local-ai'
-export * from './game-online/main'
+export * from './game-online/shared'
